requirements.txt
fastapi[all]
uvicorn
jinja2
sqlmodel
alembic
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
